A demoable version

+++ app/src/main/java/com/daniel_araujo/always_recording_microphone/PcmUtils.kt
@@ -4,8 +4,8 @@ object PcmUtils {
     /**
      * Calculates the number of bytes needed to store samples.
      */
-    fun bufferSize(milli: Int, sampleRate: Int, bytesPerSample: Int, channels: Int): Int {
-        return milli * sampleRate * bytesPerSample * channels / 1000
+    fun bufferSize(milli: Long, sampleRate: Int, bytesPerSample: Int, channels: Int): Int {
+        return (milli * sampleRate * bytesPerSample * channels / 1000).toInt()
     }
 
     /**
